pyproject.toml
[tool.poetry]
name = "hello-world"
version = "0.1.0"
description = "create an api that returns just hello world."
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pyjwt = "*"
async-exit-stack = "*"
async-generator = "*"
asyncio = "*"
bcrypt = "*"
fastapi = "*"
h11 = "*"
prisma = "*"
pydantic = "*"
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
